# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inventario Ganaderia API
  version: 1.1.0
  description: API REST para trazabilidad sanitaria, inventario y operaciones ganaderas.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /animals/import:
    post:
      tags: [Animals]
      summary: Import animals from CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Import completed
  /animals/{id}:
    get:
      tags: [Animals]
      summary: Get animal detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Animal detail
    patch:
      tags: [Animals]
      summary: Update animal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Animal updated
    delete:
      tags: [Animals]
      summary: Soft delete animal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Animal deleted
  /animals/{id}/summary:
    get:
      tags: [Animals]
      summary: Get animal timeline and stats summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Animal summary
...